+++ packages/samples/angular/src/app/atomic-angular-page/atomic-angular-page.component.ts
@@ -2,6 +2,7 @@ import {AfterViewInit, Component, ViewChild} from '@angular/core';
 import {Result, Bindings} from '@coveo/atomic-angular';
 
 @Component({
+  standalone: false,
   selector: 'app-atomic-angular-page',
   templateUrl: './atomic-angular-page.component.html',
   styleUrls: ['./atomic-angular-page.component.css'],