+++ packages/samples/headless-commerce-react/src/components/search-box/search-box.tsx
@@ -2,18 +2,36 @@ import {
   SearchBox as HeadlessSearchBox,
   InstantProducts as HeadlessInstantProducts,
   Suggestion,
+  FilterSuggestionsGenerator as HeadlessFilterSuggestionsGenerator,
+  CategoryFilterSuggestions,
+  FilterSuggestions,
+  RegularFacetSearchResult,
+  CategoryFacetSearchResult,
 } from '@coveo/headless/commerce';
 import {useEffect, useRef, useState} from 'react';
+import FilterSuggestionsGenerator from '../filter-suggestions/filter-suggestions-generator.js';
 import InstantProducts from '../instant-products/instant-products.js';
 
 interface ISearchBoxProps {
   controller: HeadlessSearchBox;
   instantProductsController: HeadlessInstantProducts;
+  filterSuggestionsGeneratorController: HeadlessFilterSuggestionsGenerator;
+  /* Uncomment the `legacyFieldSuggestionsGeneratorController property below and comment out the
+     `filterSuggestionsGeneratorController` property above if using legacy field suggestions */
+  //legacyFieldSuggestionsGeneratorController: FieldSuggestionsGenerator;
   navigate: (pathName: string) => void;
 }
 
 export default function SearchBox(props: ISearchBoxProps) {
-  const {controller, instantProductsController, navigate} = props;
+  const {
+    controller,
+    instantProductsController,
+    filterSuggestionsGeneratorController,
+    /* Uncomment the `legacyFieldSuggestionsGeneratorController property below and comment out the
+     `filterSuggestionsGeneratorController` property above if using legacy field suggestions */
+    //legacyFieldSuggestionsGeneratorController,
+    navigate,
+  } = props;
 
   const [state, setState] = useState(controller.state);
   const [isDropdownVisible, setIsDropdownVisible] = useState(false);
@@ -23,6 +41,36 @@ export default function SearchBox(props: ISearchBoxProps) {
   useEffect(() => {
     controller.subscribe(() => setState({...controller.state}));
   }, [controller]);
+
+  const fetchFilterSuggestions = (value: string) => {
+    for (const filterSuggestions of filterSuggestionsGeneratorController.filterSuggestions) {
+      filterSuggestions.updateQuery(value);
+    }
+  };
+
+  /* Uncomment the `fetchLegacyFieldSuggestions` function below and comment out the `fetchFilterSuggestions` function
+     above if using legacy field suggestions. */
+
+  // const fetchLegacyFieldSuggestions = (value: string) => {
+  //   for (const legacyFieldSuggestions of legacyFieldSuggestionsGeneratorController.fieldSuggestions) {
+  //     legacyFieldSuggestions.updateText(value);
+  //   }
+  // };
+
+  const clearFilterSuggestions = () => {
+    for (const filterSuggestions of filterSuggestionsGeneratorController.filterSuggestions) {
+      filterSuggestions.clear();
+    }
+  };
+
+  /* Uncomment the `clearLegacyFieldSuggestions` function below and comment out `clearFilterSuggestions` function above
+     if using legacy field suggestions. */
+
+  // const clearLegacyFieldSuggestions = () => {
+  //   for (const legacyFieldSuggestions of legacyFieldSuggestionsGeneratorController.fieldSuggestions) {
+  //     legacyFieldSuggestions.clear();
+  //   }
+  // };
 
   const focusSearchBoxInput = () => {
     searchInputRef.current!.focus();
@@ -44,7 +92,9 @@ export default function SearchBox(props: ISearchBoxProps) {
     }
 
     controller.updateText(e.target.value);
+    controller.showSuggestions();
     instantProductsController.updateQuery(e.target.value);
+    fetchFilterSuggestions(e.target.value);
     showDropdown();
   };
 
@@ -60,6 +110,7 @@ export default function SearchBox(props: ISearchBoxProps) {
         if (state.value !== '') {
           controller.clear();
           instantProductsController.updateQuery(state.value);
+          clearFilterSuggestions();
           break;
         }
         break;
@@ -77,6 +128,7 @@ export default function SearchBox(props: ISearchBoxProps) {
     hideDropdown();
     controller.clear();
     instantProductsController.updateQuery(state.value);
+    clearFilterSuggestions();
   };
 
   const onClickSearchBoxSubmit = () => {
@@ -87,6 +139,7 @@ export default function SearchBox(props: ISearchBoxProps) {
 
   const onFocusSuggestion = (suggestion: Suggestion) => {
     instantProductsController.updateQuery(suggestion.rawValue);
+    fetchFilterSuggestions(suggestion.rawValue);
   };
 
   const onSelectSuggestion = (suggestion: Suggestion) => {
@@ -120,11 +173,40 @@ export default function SearchBox(props: ISearchBoxProps) {
           </div>
         )}
 
-        <div className="InstantProducts column">
+        <div className="InstantProducts column small">
           <InstantProducts
             controller={instantProductsController}
             navigate={navigate}
           />
+        </div>
+
+        <div className="FilterSuggestions column small">
+          <FilterSuggestionsGenerator
+            controller={filterSuggestionsGeneratorController}
+            onClickFilterSuggestion={(
+              controller: FilterSuggestions | CategoryFilterSuggestions,
+              value: RegularFacetSearchResult | CategoryFacetSearchResult
+            ) => {
+              hideDropdown();
+              controller.type === 'hierarchical'
+                ? controller.select(value as CategoryFacetSearchResult)
+                : controller.select(value as RegularFacetSearchResult);
+            }}
+          />
+
+          {/* Uncomment the LegacyFieldSuggestionsGenerator component below and comment out the
+              FieldSuggestionsGenerator component above if using legacy field suggestions */}
+
+          {/* <LegacyFieldSuggestionsGenerator
+            controller={legacyFieldSuggestionsGeneratorController}
+            onClickLegacyFieldSuggestion={(
+              controller: CategoryFieldSuggestions,
+              value: CategoryFacetSearchResult
+            ) => {
+              hideDropdown();
+              controller.select(value);
+            }}
+          /> */}
         </div>
       </div>
     );
@@ -135,7 +217,7 @@ export default function SearchBox(props: ISearchBoxProps) {
       <input
         className="SearchBoxInput"
         id="search-box"
-        onChange={onSearchBoxInputChange}
+        onInput={onSearchBoxInputChange}
         onKeyDown={onSearchBoxInputKeyDown}
         ref={searchInputRef}
         value={state.value}