+++ packages/headless/CHANGELOG.md
@@ -1,3 +1,25 @@
+## <small>3.19.3 (2025-03-19)</small>
+
+- fix(generated answer): fix citation duplication issue in generated answer (#5053) ([09ca081](https://github.com/coveo/ui-kit/commits/09ca081)), closes [#5053](https://github.com/coveo/ui-kit/issues/5053)
+- chore: split release:phase1 & release:phase3 in two phases (#5042) ([e00bd41](https://github.com/coveo/ui-kit/commits/e00bd41)), closes [#5042](https://github.com/coveo/ui-kit/issues/5042)
+
+## <small>3.19.2 (2025-03-12)</small>
+
+- fix(headless): fix recommendation documentOpen event triggering an error for EP (#5050) ([0bb6ae6](https://github.com/coveo/ui-kit/commits/0bb6ae6)), closes [#5050](https://github.com/coveo/ui-kit/issues/5050)
+- chore: build working again on windows (#4954) ([d406941](https://github.com/coveo/ui-kit/commits/d406941)), closes [#4954](https://github.com/coveo/ui-kit/issues/4954)
+
+## <small>3.19.1 (2025-03-06)</small>
+
+- fix(deps): update all dependencies j:kit-282 (#5032) ([38b18f5](https://github.com/coveo/ui-kit/commits/38b18f5)), closes [#5032](https://github.com/coveo/ui-kit/issues/5032)
+- fix(deps): update dependency @reduxjs/toolkit to v2.6.0 j:kit-282 (#5033) ([30fb327](https://github.com/coveo/ui-kit/commits/30fb327)), closes [#5033](https://github.com/coveo/ui-kit/issues/5033)
+- chore(deps): update dependency typedoc to v0.27.9 j:kit-282 (#5030) ([f8d3d0e](https://github.com/coveo/ui-kit/commits/f8d3d0e)), closes [#5030](https://github.com/coveo/ui-kit/issues/5030)
+- feat(headless SSR commerce): Allow setting analytics 'capture' property through navigator provider ( ([25495db](https://github.com/coveo/ui-kit/commits/25495db)), closes [#4995](https://github.com/coveo/ui-kit/issues/4995)
+- feat(headless): added analytics section to fetch document suggestion call (#5011) ([6c07d5f](https://github.com/coveo/ui-kit/commits/6c07d5f)), closes [#5011](https://github.com/coveo/ui-kit/issues/5011)
+
+## 3.18.0 (2025-02-27)
+
+- feat(headless commerce): re-implement field suggestions as faceted search (#5005) ([dec06a7](https://github.com/coveo/ui-kit/commits/dec06a7)), closes [#5005](https://github.com/coveo/ui-kit/issues/5005)
+
 ## 3.17.0 (2025-02-26)
 
 - chore: update vite dependencies (#5010) ([b1f8764](https://github.com/coveo/ui-kit/commits/b1f8764)), closes [#5010](https://github.com/coveo/ui-kit/issues/5010)